requirements.txt

Django
django-cors-headers
django-filter
django-rest-swagger
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
